+++ migrations/sql/V2018.04.13.1412__UpdateBurdenEstimateStochasticForeignTable.sql
@@ -7,7 +7,7 @@ CREATE FOREIGN TABLE burden_estimate_stochastic (
   country             SMALLINT  NOT NULL,
   year                SMALLINT  NOT NULL,
   burden_outcome      SMALLINT  NOT NULL,
-  value               NUMERIC,
+  value               REAL,
   age                 SMALLINT
 )
 SERVER montagu_db_annex
